+++ src/public_generic_consumer/c_calling.rs
@@ -157,6 +157,7 @@ mod render;
 
 mod authenticated;
 pub use authenticated::{
+    generate_authenticated_allocating_calling_consumer_v1,
     generate_authenticated_identity_calling_consumer_v1,
     generate_authenticated_moves_calling_consumer_v1,
 };